requirements.txt
streamlit==1.28.1
openai==1.3.7
python-docx==0.8.11
PyPDF2==3.0.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.0.1
reportlab==4.0.4
